Add bridge attach and detach
Add the concept of bridge attachment and detachment to the
BridgeManager.  Attachment is the same as the old register method with
the change that events are addded before register is called on the
bridge itself.  Detachment is done with the cooperation of the bridge in
question, by calling it's deregister method and waiting for it to send
out a bridge_detach event.  This is necessary in order to syncronize the
action between the bridge and the bridge manager.