Rework dispatch
Rework the way event dispatch works in BrideManager.  Add a translate
stage to event dispatch that allows for modifying events, and preventing
them from being sent out and treat the manager itself as a bridge to
sent events to.

Rename the Bridge.dispatch function, which should only be called by the
BridgeManager.  Rename BaseBridge._dispatch which does something
completly different.