Refactor event system
Rename BaseEvent to Event and include a target in the event.  Special
targets for sending events to all bridges, all users, manager, etc, is
contained as class variables in the Target class.

With the inclusion of a target parameter in the event, the distinction
of who sends the event is not needed to be placed in the event name
itself, and it greatly simplifies the handling for events sent by
bridges and users to and from other bridges and users.