Make console commands local
Change the handling of commands in the console to execute commands from
ConsoleBridge instead of sending them to the bridge manager.  Add bridge
command for sending a command to a particular bridge, a shortcut manager
command for the bridge manager case and a shortcut shutdown command for
closing down the whole bridge from the console.