Add channel concept
Add a concept of channels where a bridge joins channels it wants to
bridge and receives updates for those channels when doing so.

Restructure the way the event system handles users joining and leaving.
The new system separates a bridge signaling a user joining/leaving a
channel from a broadcast that a user has joined/left a channel.  This
makes it easier to manage.