Handle channels and users in console
Add commands for joining and leaving channels in the ConsoleBridge, and
print out when the console is being added and removed from a channel.

Add code printing when users are added/updated/removed from channels the
ConsoleBridge is joined to.  And amend the name function to give name to
channels and users in channels the console bridge is part of.