Fix channel_join event leaking reference
Upon handling the channel_join event, the Bridge manager would send out
a channel_add event in response with a mutable reference to the
channel.users dict.  This would in turn lead to Bridges directly
modifying the managers list of users in channels.  Fix by passing a copy
of the users array to Bridges.