Remove Event.is_target
In bridges, only the id of targets are ever exposed through events,
which means that the bridges themselves will only be able to use
is_target on their own objects.  Unfortunately, this isn't really much
useful in practice, especially since bridges are unable to check if the
type of a target is a Bridge/Channel/User/Target.

Remove Event.is_targe due to it being virtually useless, and replace
it's only used function of checking if the target is a special Target
type with a custom __eq__ method in _TargetType.  This makes it possible
to compare the target/source id directly with a Target.