Remove mixin.py
The only usage of mixin.py was to support class tags for Manager,
Bridge, Channel, and User classes in order for it to be possible for
Event.is_target to tell which class is what.  Since is_target have been
removed this is no longer needed.

Remove mixing.py and references to the tag classes.