Track channels and users in BaseBridge
Implement the tracking of channels that are added and deleted as well as
users that join and leave the channels themselves.  This is something
all bridges more or less have to do in order to function.

Also implement hooks for these events that provides convenient User and
Channel objects with the information that relayed in these events.