Add temporary get_user() method to BaseBridge
Add helper function BaseBridge.get_user() for getting the user object
corresponding to a user with the passed id.  This is a temorary
workaround to the problem of there not being any good mapping of user id
to user object (the current setup has a user object per channel it is
in).