Add IRC bridge
Implement an IRC bridge that add an imposter bot for every user that is
added to the channel from other bridges.  PM's are properly relayed
through the bridge as expected.

Note that this perverse abuse of users joining the server is likely
going to be stopped by rate limiting and abuse filters.  Special
settings in the IRC server may be required for it to operate properly.