Dampen joins and leaves from Discord
On the mobile app for Discord, it sets the status as offline everytime
you leave the app or set the phone in standby.  This causes excessive
join and leave messages in chat platforms that show them by default.
Add a configurable timeout in which users are kept online from Discord
after they have left the server, so that there's no join/leave output
every time they check on Discord.

This is implemented by putting users that leave in a leave queue with a
timestamp, and having a separate thread periodically checking for users
that have timed out.  To avoid race conditions, a lock is introduced for
actions that require checking the current list of users.