Translate mentions over the bridge
Add a translation of mentions in the discord and irc bridges.  When a
text string of the format "@username" is sent over the bridge, it will
be translated to special mention text of the format "<[@user_id]>".

Discord translates @name mentions to a special mention of the form
"<@discord_id>", which pings the mentioned user on his client.  This is
then translated by the Discord bridge to the internal format
"<[@user_id]>".  The reverse translation is also done by the Discord
bridge when it receives a text.  In the case were the mentioned user is
in the Discord chatroom, it will turn into the special mention syntax
recognized by Discord.  Otherwise it will be translated into a plain
"@name" form.

For IRC, the translation is similar to Discord, excet that there is no
special mention syntax.  Instead "@nick" is treated as a mention, and
will be converted into the special mention if a corresponding user with
the nick is found.  The reverse translation turns the user_id into the
nickname of the user mentioned, or the nickname of the user bot that was
mentioned.