Add reconnection logic to Discord bridge
The current version of Discord.py does not attempt to reconnect in the
case of a lost connection or if the internet connection drops,
frequently raising exceptions in this case.  Some internals are also
closed when using the normal Client.run, client.start, and
client.connect methods, making reconnection difficult.

Implement reconnection by catching errors in Client.login and retrying
after 10 seconds.  And looping over a custom implementation of
Client.connect, catching errors that may be raised, until the Client is
closed.