Properly notify unexpected exception on login
During login of the Discord bot, an unexpected exception may occur which
would cause the thread to exit without notifying of it.  Add a similar
except clause to the login part of the discord bot as is done with the
main running part.  Also clarify that it should exit the function rather
than just break out of the loop in the event of an unexpected exception.