Accept new Discord mention syntax
An api change in Discord brought an optional exclamation mark in the
metion syntax, signaling that the mentioned user has a nickname.
Currently the reverse translation is not done, as the data needed to
determine whether to put the sign there or not is not available at the
location it's done.